requirements.txt
# AI Data Analyst Agent - Core Dependencies
# Install with: pip install -r requirements.txt

# Data Processing
pandas>=1.3.0
numpy>=1.20.0

# Statistical Analysis
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# CLI Framework
click>=8.0.0

# Excel Support
openpyxl>=3.7.0

# YAML Configuration
PyYAML>=5.4.0


# ============================================================
# DEVELOPMENT DEPENDENCIES (optional)
# ============================================================
# Install with: pip install -r requirements.txt && pip install -r requirements-dev.txt

# Development Requirements (save as requirements-dev.txt):
# pytest>=6.2.0
# pytest-cov>=2.12.0
# black>=21.0
# flake8>=3.9.0
# mypy>=0.910
# ipython>=7.0
# jupyter>=1.0
# sphinx>=4.0.0
# sphinx-rtd-theme>=1.0.0
